requirements.txt
fastmcp>=0.4.0
supabase>=2.0.0
sentence-transformers>=2.2.0
requests>=2.28.0
stripe>=7.0.0
python-dotenv>=1.0.0
